refactor(eve): split remote subagent startup
